Create an extension function for each existing method in RealmManager
