+++ app/src/main/kotlin/com/mgaetan89/showsrage/extension/RealmExtension.kt
@@ -7,12 +7,17 @@ import com.mgaetan89.showsrage.model.LogLevel
 import com.mgaetan89.showsrage.model.OmDbEpisode
 import com.mgaetan89.showsrage.model.Quality
 import com.mgaetan89.showsrage.model.RealmShowStat
+import com.mgaetan89.showsrage.model.RealmString
 import com.mgaetan89.showsrage.model.RootDir
 import com.mgaetan89.showsrage.model.Schedule
+import com.mgaetan89.showsrage.model.Serie
 import com.mgaetan89.showsrage.model.Show
+import com.mgaetan89.showsrage.model.ShowStat
 import com.mgaetan89.showsrage.model.ShowWidget
+import com.mgaetan89.showsrage.model.ShowsStat
 import io.realm.Realm
 import io.realm.RealmChangeListener
+import io.realm.RealmList
 import io.realm.RealmResults
 import io.realm.Sort
 
@@ -62,6 +67,15 @@ fun Realm.deleteShow(indexerId: Int) {
     }
 }
 
+fun Realm.deleteShowWidget(widgetId: Int) {
+    this.executeTransaction {
+        it.where(ShowWidget::class.java)
+                .equalTo("widgetId", widgetId)
+                .findAll()
+                .deleteAllFromRealm()
+    }
+}
+
 fun Realm.getEpisode(episodeId: String, listener: RealmChangeListener<Episode>?): Episode? {
     val query = this.where(Episode::class.java)
             .equalTo("id", episodeId)
@@ -85,7 +99,7 @@ fun Realm.getEpisodes(episodeId: String, listener: RealmChangeListener<RealmResu
     return episodes
 }
 
-fun Realm.getEpisodes(indexerId: Int, season: Int, reversedOrder: Boolean, listener: RealmChangeListener<RealmResults<Episode>>?): RealmResults<Episode> {
+fun Realm.getEpisodes(indexerId: Int, season: Int, reversedOrder: Boolean, listener: RealmChangeListener<RealmResults<Episode>>): RealmResults<Episode> {
     val episodes = this.where(Episode::class.java)
             .equalTo("indexerId", indexerId)
             .equalTo("season", season)
@@ -103,12 +117,8 @@ fun Realm.getHistory(listener: RealmChangeListener<RealmResults<History>>): Real
     return history
 }
 
-fun Realm.getLogs(logLevel: LogLevel, listener: RealmChangeListener<RealmResults<LogEntry>>): RealmResults<LogEntry> {
-    return this.getLogs(logLevel, null, listener)
-}
-
 fun Realm.getLogs(logLevel: LogLevel, groups: Array<String>?, listener: RealmChangeListener<RealmResults<LogEntry>>): RealmResults<LogEntry> {
-    val logLevels = getLogLevels(logLevel)
+    val logLevels = logLevel.logLevels
     val query = this.where(LogEntry::class.java)
 
     if (groups != null && groups.isNotEmpty()) {
@@ -128,7 +138,7 @@ fun Realm.getLogs(logLevel: LogLevel, groups: Array<String>?, listener: RealmCha
 fun Realm.getLogsGroup(): List<String> {
     return this.where(LogEntry::class.java)
             .distinct("group")
-            .map { it.group }
+            .map(LogEntry::group)
             .filterNotNull()
             .sorted()
 }
@@ -138,11 +148,257 @@ fun Realm.getRootDirs(): RealmResults<RootDir> {
             .findAll()
 }
 
-private fun getLogLevels(logLevel: LogLevel): Array<String> {
-    return when (logLevel) {
-        LogLevel.DEBUG -> arrayOf("DEBUG", "ERROR", "INFO", "WARNING")
-        LogLevel.ERROR -> arrayOf("ERROR")
-        LogLevel.INFO -> arrayOf("ERROR", "INFO", "WARNING")
-        LogLevel.WARNING -> arrayOf("ERROR", "WARNING")
+fun Realm.getSchedule(section: String, listener: RealmChangeListener<RealmResults<Schedule>>): RealmResults<Schedule> {
+    val schedule = this.where(Schedule::class.java)
+            .equalTo("section", section)
+            .findAllSortedAsync("airDate")
+    schedule.addChangeListener(listener)
+
+    return schedule
+}
+
+fun Realm.getScheduleSections(): List<String> {
+    return this.where(Schedule::class.java)
+            .distinct("section")
+            .map(Schedule::section)
+}
+
+fun Realm.getSeries(imdbId: String, listener: RealmChangeListener<RealmResults<Serie>>): RealmResults<Serie> {
+    val series = this.where(Serie::class.java)
+            .equalTo("imdbId", imdbId)
+            .findAllAsync()
+    series.addChangeListener(listener)
+
+    return series
+}
+
+fun Realm.getShow(indexerId: Int, listener: RealmChangeListener<Show>? = null): Show? {
+    val query = this.where(Show::class.java)
+            .equalTo("indexerId", indexerId)
+
+    if (listener == null) {
+        return query.findFirst()
     }
+
+    val show = query.findFirstAsync()
+    show.addChangeListener(listener)
+
+    return show
+}
+
+fun Realm.getShows(anime: Boolean?, listener: RealmChangeListener<RealmResults<Show>>?): RealmResults<Show>? {
+    val query = this.where(Show::class.java)
+
+    if (anime != null) {
+        query.equalTo("anime", if (anime) 1 else 0)
+    }
+
+    if (listener == null) {
+        return query.findAllSorted("showName")
+    }
+
+    val shows = query.findAllSortedAsync("showName")
+    shows.addChangeListener(listener)
+
+    return shows
+}
+
+fun Realm.getShowsStats(listener: RealmChangeListener<RealmResults<ShowsStat>>): RealmResults<ShowsStat> {
+    // There might be no data yet. So we request all data to be sure to always received a valid object.
+    val stats = this.where(ShowsStat::class.java)
+            .findAllAsync()
+    stats.addChangeListener(listener)
+
+    return stats
+}
+
+fun Realm.getShowStats(indexerId: Int): RealmShowStat? {
+    return this.where(RealmShowStat::class.java)
+            .equalTo("indexerId", indexerId)
+            .findFirst()
+}
+
+fun Realm.getShowWidget(widgetId: Int): ShowWidget? {
+    return this.where(ShowWidget::class.java)
+            .equalTo("widgetId", widgetId)
+            .findFirst()
+}
+
+fun Realm.saveEpisode(episode: Episode, indexerId: Int, season: Int, episodeNumber: Int) {
+    this.executeTransaction {
+        it.prepareEpisodeForSaving(episode, indexerId, season, episodeNumber)
+
+        it.copyToRealmOrUpdate(episode)
+    }
+}
+
+fun Realm.saveEpisode(episode: OmDbEpisode) {
+    this.executeTransaction {
+        prepareEpisodeForSaving(episode)
+
+        it.copyToRealmOrUpdate(episode)
+    }
+}
+
+fun Realm.saveEpisodes(episodes: List<Episode>, indexerId: Int, season: Int) {
+    this.executeTransaction {
+        episodes.forEach { episode ->
+            it.prepareEpisodeForSaving(episode, indexerId, season, episode.number)
+        }
+
+        it.copyToRealmOrUpdate(episodes)
+    }
+}
+
+fun Realm.saveHistory(histories: List<History>) {
+    this.clearHistory()
+
+    this.executeTransaction {
+        histories.forEach(::prepareHistoryForSaving)
+
+        it.copyToRealmOrUpdate(histories)
+    }
+}
+
+fun Realm.saveLogs(logLevel: LogLevel, logs: List<LogEntry>) {
+    this.executeTransaction {
+        // Remove all existing logs for the level we are about to save
+        it.where(LogEntry::class.java)
+                .equalTo("errorType", logLevel.name)
+                .findAll()
+                .deleteAllFromRealm()
+
+        // Save the new logs in the database
+        it.copyToRealm(logs)
+    }
+}
+
+fun Realm.saveRootDirs(rootDirs: List<RootDir>) {
+    this.executeTransaction {
+        it.delete(RootDir::class.java)
+        it.copyToRealmOrUpdate(rootDirs)
+    }
+}
+
+fun Realm.saveSchedules(section: String, schedules: List<Schedule>) {
+    this.executeTransaction {
+        schedules.forEach {
+            prepareScheduleForSaving(it, section)
+        }
+
+        it.copyToRealmOrUpdate(schedules)
+    }
+}
+
+fun Realm.saveSerie(serie: Serie) {
+    this.executeTransaction {
+        it.copyToRealmOrUpdate(serie)
+    }
+}
+
+fun Realm.saveShow(show: Show) {
+    this.executeTransaction {
+        it.prepareShowForSaving(show)
+
+        it.copyToRealmOrUpdate(show)
+    }
+}
+
+fun Realm.saveShows(shows: List<Show>) {
+    this.executeTransaction {
+        shows.forEach { show ->
+            it.prepareShowForSaving(show)
+        }
+
+        it.copyToRealmOrUpdate(shows)
+    }
+
+    // Remove information about shows that might have been removed
+    val savedShows = this.getShows(null, null) ?: return
+    val removedIndexerIds = savedShows.map(Show::indexerId) - shows.map(Show::indexerId)
+
+    removedIndexerIds.forEach {
+        // deleteShow has its own transaction, so we need to run this outside of the above transaction
+        this.deleteShow(it)
+    }
+}
+
+fun Realm.saveShowsStat(stat: ShowsStat) {
+    this.executeTransaction {
+        it.delete(ShowsStat::class.java)
+        it.copyToRealm(stat)
+    }
+}
+
+fun Realm.saveShowStat(stat: ShowStat, indexerId: Int): RealmShowStat {
+    val realmStat = getRealmShowStat(stat, indexerId)
+
+    this.executeTransaction {
+        it.copyToRealmOrUpdate(realmStat)
+    }
+
+    return realmStat
+}
+
+fun Realm.saveShowWidget(showWidget: ShowWidget) {
+    this.executeTransaction {
+        it.copyToRealmOrUpdate(showWidget)
+    }
+}
+
+private fun Realm.prepareEpisodeForSaving(episode: Episode, indexerId: Int, season: Int, episodeNumber: Int) {
+    val id = Episode.buildId(indexerId, season, episodeNumber)
+    val savedEpisode = this.getEpisode(id, null)
+
+    episode.description = if (episode.description.isNullOrEmpty()) savedEpisode?.description else episode.description
+    episode.fileSizeHuman = if (episode.fileSizeHuman.isNullOrEmpty()) savedEpisode?.fileSizeHuman else episode.fileSizeHuman
+    episode.id = id
+    episode.indexerId = indexerId
+    episode.number = episodeNumber
+    episode.season = season
+}
+
+private fun Realm.prepareShowForSaving(show: Show) {
+    val savedShow = this.getShow(show.indexerId)
+
+    show.airs = if (show.airs.isNullOrEmpty()) savedShow?.airs else show.airs
+    show.genre = RealmList<RealmString>().apply {
+        this.addAll((if (show.genre?.isEmpty() ?: true) savedShow?.genre else show.genre) ?: emptyList())
+    }
+    show.imdbId = if (show.imdbId.isNullOrEmpty()) savedShow?.imdbId else show.imdbId
+    show.location = if (show.location.isNullOrEmpty()) savedShow?.location else show.location
+    show.qualityDetails = Quality().apply {
+        this.archive = RealmList<RealmString>().apply {
+            this.addAll((if (show.qualityDetails?.archive?.isEmpty() ?: true) savedShow?.qualityDetails?.archive else show.qualityDetails?.archive) ?: emptyList())
+        }
+        this.indexerId = show.indexerId
+        this.initial = RealmList<RealmString>().apply {
+            this.addAll((if (show.qualityDetails?.initial?.isEmpty() ?: true) savedShow?.qualityDetails?.initial else show.qualityDetails?.initial) ?: emptyList())
+        }
+    }
+    show.seasonList = RealmList<RealmString>().apply {
+        this.addAll((if (show.seasonList?.isEmpty() ?: true) savedShow?.seasonList else show.seasonList) ?: emptyList())
+    }
+}
+
+private fun getRealmShowStat(stat: ShowStat, indexerId: Int): RealmShowStat {
+    return RealmShowStat().apply {
+        this.downloaded = stat.getTotalDone()
+        this.episodesCount = stat.total
+        this.indexerId = indexerId
+        this.snatched = stat.getTotalPending()
+    }
+}
+
+private fun prepareEpisodeForSaving(episode: OmDbEpisode) {
+    episode.id = OmDbEpisode.buildId(episode.seriesId ?: "", episode.season ?: "", episode.episode ?: "")
+}
+
+private fun prepareHistoryForSaving(history: History) {
+    history.id = "${history.date}_${history.status}_${history.indexerId}_${history.season}_${history.episode}"
+}
+
+private fun prepareScheduleForSaving(schedule: Schedule, section: String) {
+    schedule.id = "${schedule.indexerId}_${schedule.season}_${schedule.episode}"
+    schedule.section = section
 }
