+++ app/src/main/kotlin/com/mgaetan89/showsrage/model/LogLevel.kt
@@ -2,11 +2,11 @@ package com.mgaetan89.showsrage.model
 
 import java.util.*
 
-enum class LogLevel {
-    DEBUG,
-    ERROR,
-    INFO,
-    WARNING;
+enum class LogLevel(val logLevels: Array<String>) {
+    DEBUG(arrayOf("DEBUG", "ERROR", "INFO", "WARNING")),
+    ERROR(arrayOf("ERROR")),
+    INFO(arrayOf("ERROR", "INFO", "WARNING")),
+    WARNING(arrayOf("ERROR", "WARNING"));
 
     override fun toString(): String {
         return super.toString().toLowerCase(Locale.getDefault())
